Factor merge_dict out from edalizer into utils
This function is helpful in other scenarios as well.
