+++ fusesoc/edalizer.py
@@ -5,6 +5,7 @@ import shutil
 
 from fusesoc import utils
 from fusesoc.vlnv import Vlnv
+from fusesoc.utils import merge_dict
 
 logger = logging.getLogger(__name__)
 
@@ -38,16 +39,6 @@ class Edalizer:
             os.makedirs(work_root)
 
         logger.debug("Building EDA API")
-
-        def merge_dict(d1, d2):
-            for key, value in d2.items():
-                if isinstance(value, dict):
-                    d1[key] = merge_dict(d1.get(key, {}), value)
-                elif isinstance(value, list):
-                    d1[key] = d1.get(key, []) + value
-                else:
-                    d1[key] = value
-            return d1
 
         generators = {}
